requirements.txt
snowflake-snowpark-python
pandas
pillow
pyarrow

openai
requests
simplejson
langchain>=0.1.13
langchain-community
langchain_openai
python-dotenv>=1.0.1
snowflake>=0.5.1
snowflake-snowpark-python
flask>=3.0.2
plotly>=5.20.0
streamlit-authenticator
tabulate>=0.9.0
tiktoken
faiss-cpu
pyyaml

streamlit-extras==0.4.0
streamlit_option_menu==0.3.2
streamlit==1.35.0

flask
gunicorn
snowflake-connector-python
